Add shared login helper for system specs and handle notifications
